+++ src/modules/component-builder.ts
@@ -4,7 +4,10 @@ import { ComponentProperties } from "./jsx";
 function buildFunctionalComponent(name, code, attributes: ComponentProperties) {
   const props = new Set([
     ...attributes.argumentProps,
-    ...attributes.componentMembers
+    ...attributes.componentMembers,
+    ...attributes.memberProps,
+    ...attributes.state
+
   ]);
   return `
     function ${name}({${Array.from(props).join(", ")}}) {
@@ -32,9 +35,5 @@ function buildStatefulComponent(name, code, attributes: ComponentProperties) {
 }
 
 export function buildComponent(name, code, attributes) {
-  if (attributes.memberProps.size || attributes.state.size) {
-    return buildStatefulComponent(name, code, attributes);
-  } else {
     return buildFunctionalComponent(name, code, attributes);
-  }
 }
