+++ src/parsing.ts
@@ -16,6 +16,14 @@ export const codeToAst = code =>
     ...parsingOptions
   });
 
+  export const jsxToAst = code => {
+    try {
+      return codeToAst(code);
+    } catch (e) {
+      return codeToAst(`<>${code}</>`);
+    }
+  };
+
 export const templateToAst = code => template.ast(code, parsingOptions);
 
 export function getIdentifier(code) {
